requirements.txt
BeautifulSoup
django-taggit
django-markup
South